src/manga-day/requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
bs4==0.0.2
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fake-headers==1.0.2
fastapi==0.128.0
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
html5lib==1.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
loguru==0.7.3
lxml==6.0.2
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
packaging==26.0
patchright==1.58.0
pillow==12.1.0
pluggy==1.6.0
propcache==0.4.1
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
webencodings==0.5.1
websockets==16.0
win32_setctime==1.2.0
yarl==1.22.0
psycopg2-binary==2.9.11
alembic==1.18.3
python-dotenv==1.2.1